Add interval and dayprefix to schedule
Add interval and day prefix to regular schedule, parsed from free text schedule.

Interval is how frequently it occurs and relates to the freq column. e.g. if fortnightly then interval = 2 and freq = WEEKLY

DayPrefix is a prefix for the day field to show which day of the month it falls on, e.g. if 3rd Monday of every month then byday = 3MO and freq = MONTHLY